added orders

+++ DRHC/Controllers/TestimonialController.cs
@@ -33,20 +33,21 @@ namespace DRHC.Controllers
 
         public async Task<int> GetUserDetails(ApplicationUser user)
         {
+
             if (user == null) return 0;
           
             if (user.AdminID == null) return 1; //User is not admin author
-            else return 2;
+            if (user.AdminID != null) return 2; //User is not admin author
+
+            else return -1;
             
-            return -1;//something went wrong
         }
 
 
         //for user display
-        public ActionResult userview()
+        public ActionResult _userview()
         {
-            return View( "~/Views/Testimonial/userview.cshtml", db.Testimonials);
-            //return View(db.Testimonials);
+            return View(db.Testimonials);
         }
 
 
@@ -60,25 +61,31 @@ namespace DRHC.Controllers
         [HttpPost]
         public ActionResult Create(string UserFName, string UserLName, string Title, string Story)
         {
-            string query = "insert into Testimonials (UserFName, UserLName, Title,Story,TestimonialStatusID) " +
+
+            if (ModelState.IsValid)
+            {
+
+                string query = "insert into Testimonials (UserFName, UserLName, Title,Story,TestimonialStatusID) " +
                 "values (@UserFName, @UserLName, @Title,@Story,@TestimonialStatusID)";
 
-            //Making deafult status of testimonial to Pending-Unpublished
-            var Ts = db.TestimonialStatuss.SingleOrDefault(ts => ts.TestimonialStatusID == 3);
-            var tsid = Ts.TestimonialStatusID;
-                
-            SqlParameter[] myparams = new SqlParameter[5];
-            myparams[0] = new SqlParameter("@UserFName", UserFName);
-            myparams[1] = new SqlParameter("@UserLName", UserLName);
-            myparams[2] = new SqlParameter("@Title", Title);
-            myparams[3] = new SqlParameter("@Story", Story);
-            myparams[4] = new SqlParameter("@TestimonialStatusID", tsid);
+                //Making deafult status of testimonial to Pending-Unpublished
+                var Ts = db.TestimonialStatuss.SingleOrDefault(ts => ts.TestimonialStatusID == 3);
+                var tsid = Ts.TestimonialStatusID;
 
-            db.Database.ExecuteSqlCommand(query, myparams);
-            List<Testimonial> t = db.Testimonials.ToList();
-            int id = (int)t.Max(testimonial => testimonial.TestimonialID);
+                SqlParameter[] myparams = new SqlParameter[5];
+                myparams[0] = new SqlParameter("@UserFName", UserFName);
+                myparams[1] = new SqlParameter("@UserLName", UserLName);
+                myparams[2] = new SqlParameter("@Title", Title);
+                myparams[3] = new SqlParameter("@Story", Story);
+                myparams[4] = new SqlParameter("@TestimonialStatusID", tsid);
 
-            return RedirectToAction("Sendletter/"+ id + "&tid=" + tsid);
+                db.Database.ExecuteSqlCommand(query, myparams);
+                List<Testimonial> t = db.Testimonials.ToList();
+                int id = (int)t.Max(testimonial => testimonial.TestimonialID);
+                return RedirectToAction("Sendletter", new { iid = id, tid = tsid });
+            }
+            else return View("Index","Home");
+
         }
 
 
@@ -87,48 +94,60 @@ namespace DRHC.Controllers
             var user = await GetCurrentUserAsync();
             var userstate = await GetUserDetails(user);
 
-           if (userstate==0){
-                return RedirectToAction("Register", "Account");}
-
-            var testimonials = await db.Testimonials.Include(t => t.TestimonialStatus).ToListAsync();
-            int count = testimonials.Count();
-            int perpage = 3;
-            int maxpage = (int)Math.Ceiling((decimal)count / perpage) - 1;
-            if (maxpage < 0) maxpage = 0;
-            if (pagenum < 0) pagenum = 0;
-            if (pagenum > maxpage) pagenum = maxpage;
-            int start = perpage * pagenum;
-            ViewData["pagenum"] = (int)pagenum;
-            ViewData["PaginationSummary"] = "";
-            if (maxpage > 0)
+            if (userstate == 2)
             {
-                ViewData["PaginationSummary"] =
-                    (pagenum + 1).ToString() + " of " +
-                    (maxpage + 1).ToString();
+
+                var testimonials = await db.Testimonials.Include(t => t.TestimonialStatus).ToListAsync();
+                int count = testimonials.Count();
+                int perpage = 3;
+                int maxpage = (int)Math.Ceiling((decimal)count / perpage) - 1;
+                if (maxpage < 0) maxpage = 0;
+                if (pagenum < 0) pagenum = 0;
+                if (pagenum > maxpage) pagenum = maxpage;
+                int start = perpage * pagenum;
+                ViewData["pagenum"] = (int)pagenum;
+                ViewData["PaginationSummary"] = "";
+                if (maxpage > 0)
+                {
+                    ViewData["PaginationSummary"] =
+                        (pagenum + 1).ToString() + " of " +
+                        (maxpage + 1).ToString();
+                }
+
+                List<Testimonial> testi = await db.Testimonials.Include(t => t.TestimonialStatus).Skip(start).Take(perpage).ToListAsync();
+                /*End of Pagination  Algorithm*/
+
+                return View(testi);
             }
-
-            List<Testimonial> testi = await db.Testimonials.Include(t => t.TestimonialStatus).Skip(start).Take(perpage).ToListAsync();
-            /*End of Pagination  Algorithm*/
-
-            return View(testi);
+            else return View("Index", "Home");
 
         }
 
+        public async Task<ActionResult> Show(int id)
+        {
+            var user = await GetCurrentUserAsync();
+            var userstate = await GetUserDetails(user);
+
+            if (userstate == 2)
+            {
+
+                Testimonial t = db.Testimonials.Include(tm => tm.TestimonialStatus)
+                    .SingleOrDefault(tm => tm.TestimonialID == id);
+
+            return View(t);
+            }
+            else return View("Index", "Home");
+
+        }
 
 
         public async Task<ActionResult> Edit(int id)
         {
             var user = await GetCurrentUserAsync();
             var userstate = await GetUserDetails(user);
-            if (userstate == 0)
+            if (userstate == 2)
             {
-                return RedirectToAction("Register", "Account");
-            }
-
-
-           
             TestimonialEdit te = new TestimonialEdit();
-            
             te.Testimonial =
                 db.Testimonials
                     .Include(t => t.TestimonialStatus)
@@ -137,6 +156,9 @@ namespace DRHC.Controllers
            
             if (te.Testimonial != null) return View(te);
             else return NotFound();
+            }
+            else return View("Index", "Home");
+
         }
 
         [HttpPost]
@@ -144,10 +166,9 @@ namespace DRHC.Controllers
         {
             var user = await GetCurrentUserAsync();
             var userstate = await GetUserDetails(user);
-            if (userstate == 0)
+            if (userstate == 2)
             {
-                return RedirectToAction("Register", "Account");
-            }
+               
             if (db.Testimonials.Find(id) == null)
             {
                 //Show error message
@@ -162,58 +183,90 @@ namespace DRHC.Controllers
             myparams[1] = new SqlParameter("@id", id);
             
             db.Database.ExecuteSqlCommand(query, myparams);
-            return RedirectToAction("Sendletter/" + id + "&tid=" + TestimonialStatusID);
+            TestimonialStatus ts = db.TestimonialStatuss.Find(TestimonialStatusID);
+            if (ts.TestimonialStatusName.Equals("Approved-Published"))
+            {
+                return RedirectToAction("Sendletter", new { iid = id, tid = TestimonialStatusID });
+            }
+            else
+            {
+                return RedirectToAction("List");
+
+            }
+            }
+            else return View("Index", "Home");
+
+
+
+
         }
 
-        
-        public ActionResult Delete(int id)
+
+        public async Task<ActionResult> Delete(int id)
         {
-            string query = "delete from Testimonials where TestimonialID = @id";
+            var user = await GetCurrentUserAsync();
+            var userstate = await GetUserDetails(user);
+            if (userstate == 2)
+            {
+
+                string query = "delete from Testimonials where TestimonialID = @id";
             db.Database.ExecuteSqlCommand(query, new SqlParameter("@id", id));
 
            
             return RedirectToAction("List");
         }
+            else return View("Index", "Home");
+
+    }
 
 
-        //sending thankyou letter to user
-        public ActionResult Sendletter(int? id , int? tid)
-        {
-            Testimonial t = db.Testimonials.Find(id);
-            TestimonialStatus ts = db.TestimonialStatuss.Find(tid);
-            string Message;
-            if (t.TestimonialStatus.Equals("Aproved-Published"))
-            {
-                Message = "Thankyou " + t.UserFName + " " + t.UserLName + " Your Testimonial is being recieved and is under revewing process.";
+    //sending thankyou letter to user
+    public async Task<ActionResult> Sendletter(int iid , int tid)
+         {
+            int id = iid;
+             Testimonial t = db.Testimonials.Find(id);
+             TestimonialStatus ts = db.TestimonialStatuss.Find(tid);
+             string Message;
+             if (ts.TestimonialStatusName.Equals("Approved-Published"))
+             {
+                 Message = "Thankyou " + t.UserFName + " " + t.UserLName + "Your Testimonial is being Published and can be seen on home page.";
             }
-            else
-            {
-                Message = "Thankyou " + t.UserFName + " " + t.UserLName + " Your Testimonial is being Published and can be seen on home page.";
+             else
+             {
+                 Message = "Thankyou " + t.UserFName + " " + t.UserLName + "  Your Testimonial is being recieved and is under revewing process.";
+
             }
 
             var msg = new MimeMessage();
-            msg.From.Add(new MailboxAddress("ram", "rohitinventor2@gmail.com"));
-            msg.To.Add(new MailboxAddress(t.UserFName, "rohitinventor2@gmail.com"));
-            msg.Subject = "Testimonial";
-            msg.Body = new TextPart("html")
+             msg.From.Add(new MailboxAddress("admin", "wdn01269796@gmail.com"));
+             msg.To.Add(new MailboxAddress(t.UserFName, "rohitinventor1@gmail.com"));
+             msg.Subject = "Testimonial";
+             msg.Body = new TextPart("html")
+             {
+                 Text = Message
+             };
+
+
+             using (var client = new MailKit.Net.Smtp.SmtpClient())
+             {
+                 client.Connect("smtp.gmail.com", 587, false);
+                 client.Authenticate("wdn01269796@gmail.com", "mailtest1234");
+                 client.Send(msg);
+                 client.Disconnect(true);
+
+
+
+             }
+
+            var user = await GetCurrentUserAsync();
+            var userstate = await GetUserDetails(user);
+            if (userstate == 2)
             {
-                Text = Message
-            };
-
-
-            using (var client = new MailKit.Net.Smtp.SmtpClient())
-            {
-                client.Connect("smtp.gmail.com", 587, false);
-                client.Authenticate("rohitinventor2@gmail.com", "ramkisan");
-                client.Send(msg);
-                client.Disconnect(true);
-
-
+               return RedirectToAction("List");
 
             }
+            else return RedirectToAction("Index", "Home");
 
-
-            return RedirectToAction("List");
 
         }
